Factor out the networking code.
