+++ insights/tests/client/connection/test_branch_info.py
@@ -2,10 +2,9 @@ from insights.client.connection import InsightsConnection
 from mock.mock import Mock, patch
 
 
-@patch("insights.client.connection.InsightsConnection._init_session")
-@patch("insights.client.connection.InsightsConnection.get_proxies")
+@patch("insights.client.connection.InsightsConnection.new_session")
 @patch("insights.client.connection.constants.cached_branch_info", "/tmp/insights-test-cached-branchinfo")
-def test_request(get_proxies, init_session):
+def test_request(new_session):
     """
     The request to get branch info is issued with correct timeout set.
     """
@@ -14,4 +13,4 @@ def test_request(get_proxies, init_session):
     connection = InsightsConnection(config)
     connection.branch_info()
 
-    init_session.return_value.get.assert_called_once_with(config.branch_info_url, timeout=config.http_timeout)
+    new_session.return_value.get.assert_called_once_with(config.branch_info_url)
