+++ insights/tests/client/connection/test_init_session.py
@@ -1,31 +1,29 @@
-from insights.client.connection import InsightsConnection
-from mock.mock import Mock, patch
+from insights.client.session import InsightsSession
+from mock.mock import MagicMock, Mock, patch
 
 
-@patch("insights.client.connection.requests.Session")
-@patch("insights.client.connection.InsightsConnection.__init__", return_value=None)
-def test_proxy_request(init, session):
+@patch("insights.client.session.InsightsSession.request", MagicMock())
+@patch("insights.client.session.InsightsSession._set_proxy_authorization",
+        MagicMock())
+def test_proxy_request():
     """
-    If proxy authentication is enabled, the failing hacky request is issued with correct timeout set.
+    If proxy authentication is enabled, the failing hacky request is issued
+    with correct timeout set.
     """
     config = Mock()
 
-    # The constructor is bypassed, because it itself runs the _init_session method.
-    connection = InsightsConnection(None)
-    connection.config = config
-    connection.user_agent = None
-    connection.systemid = None
-    connection.authmethod = "BASIC"
-    connection.username = "some username"
-    connection.password = "some password"
-    connection.cert_verify = None
-    connection.proxies = {"https": "some proxy"}
-    connection.proxy_auth = True
+    config.user_agent = None
+    config.systemid = None
+    config.authmethod = "BASIC"
+    config.username = "some username"
+    config.password = "some password"
+    config.cert_verify = None
+    config.proxy = "https://foo:bar@www.example.com"
+    config.base_url = "cert-api.access.redhat.com/r/insights"
+    config.insecure_connection = True
+    session = InsightsSession.from_config(config)
 
-    connection._init_session()
-
-    session.return_value.request.assert_called_once_with(
+    session.request.assert_called_once_with(
         "GET",
-        "https://cert-api.access.redhat.com/r/insights",
-        timeout=config.http_timeout
+        "https://cert-api.access.redhat.com/r/insights"
     )
